Merged PR 18239: bug: Fix issue with a/b update from containers
# 🔍 Description
Recent removal of chroot logic from container broke a/b update from a container. This was unfortunately not discovered as part of the validation, as the manual steps for a/b validation from container were marked as broken. Upon further inspection, they were doing an extra stop, which made them broken.

This PR fixes both the instructions for A/B update from container and the actual A/B update from a container.

In addition, adds extra check to not return None for A/B volume selection outside of clean-install, as it can lead to data loss (volume A being overwritten, while booted from A).

# 🤔 Rationale
Fix A/B update from a container.

Related work items: #6650