Merged PR 18266: bug: Prevent trident from mounting over populated directories
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->
Trident currently allows for a mount point to have a path to a directory that could have files on it. For example, I could have a mount point path to `/var/lib/rpm` which overrides the content in the original directory. This small check allows us to not be mounting on populated directories.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->
Prevent mounting on populated directories.

# 📝 Checks
<!-- Note any manual checks you've run, base list of tests: dev-docs/manual-validation.md -->

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

- #0000

# 🗒️ Notes
<!-- Add any other relevant info -->

Related work items: #6710