Merged PR 17988: engineering: functional-tests: OS deployment is now a test
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->

- Perform the creation of the environment as a pytest test case instead of a fixture.
- Bring in the depends pytest plug-in I made for mhci-cli to add test dependencies here.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

Right now if trident fails we get a very annoying avalanche of errors because a failed fixture fails all the tests that use it. With this change, the deployment happens in a test. All the other tests are dependents and will be skipped if the VM deployment failed.

# 📝 Checks
<!-- Note any manual checks you've run -->

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

# 🗒️ Notes
<!-- Add any other relevant info -->

Example output:
![image.png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/17988/attachments/image.png)