Merged PR 18296: engineering: update the grub config at AZLA/AZLB
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->
we are still updating the grub config at EFI/BOOT instead of /EFI/AZLA or /EFI/AZLB.
This PR changes the grub config path to be modifies the boot entries grub config.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

# 📝 Checks
<!-- Note any manual checks you've run, base list of tests: dev-docs/manual-validation.md -->

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

- #0000

# 🗒️ Notes
<!-- Add any other relevant info -->

Related work items: #6540