Merged PR 18226: feature: Add support for additional files
# 🔍 Description

Adds the ability to indicate additional files in the HC that should be added during provisioning.

# 🤔 Rationale

MIC provides this feature and customers have requested it. This should be slightly easier than doing the same thing via scripts.

The functionality itself lives in the hooks module because it already had the groundwork for staging files during the prepare step and then using them during later steps.

# 📝 Checks

Manually verified a couple combinations.

Related work items: #6318