Merged PR 18290: bug: Append entries to boot order upon creation
# 🔍 Description

Certain Dell firmware seems to refuse to follow BootNext if the indicated entry is not present in the boot order list. This PR has Trident append a newly created entry to the end of the boot order list.

It also means that we can add a post-provision script to run `efibootmgr --delete-bootorder` and then the _only_ boot entry will be the one added by Trident, so we can workaround cases were BootNext is fully ignored

Related work items: #7022