Merged PR 18256: engineering: Track HostConfiguration within HostStatus
# 🔍 Description

The HostConfiguration is stored as a field within HostStatus.

# 🤔 Rationale

This lets Trident tell whether it has gotten a new HC or if the HC is unchanged. It should be helpful when splitting up Stage / Finalize because we'll know whether anything has changed between the stage call and the later finalize call.

At the moment there'll be significant duplication between what's stored in the host status and in the spec field of the host status, but that can hopefully be addressed gradually in follow-ups.

Related work items: #7047