Merged PR 18386: feature: Document Trident usage better and fix dependencies
# 🔍 Description
Adds documentation on how to use release bits, what Trident RPMs we are producing, where Trident can run and what optional dependencies need to be installed.

# 🤔 Rationale
As we are getting more customers, each with different needs, in order to scale, we need to have all these docs easily accessible.

Additionally, it adds required dependencies for provisioning into the Requires of the Trident RPM.

Related work items: #6650