Merged PR 18416: bug: set a fixed version of nextest-rs
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->

Set a fixed version of nextest compatible with rust 1.72

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

> [1:54 PM] Jonathan Behrens
> The `cargo-nextest` binary released a new version yesterday that requires `1.73`. Our pipeline [which runs on 1.72] automatically installs the latest version (because no version number is specified in the cargo install command)

Related work items: #7165