Merged PR 18419: engineering: Antivirus, wget doesnt trigger error at first attempt
# 🔍 Description
- Don't trigger error if wget fails at first attempt.
- Step continues in error.

# 🤔 Rationale
Even though if wget command succeeds in downloading after some attempts, because the first attempt failed, it will return a non-zero status, which will trigger an error in the pipeline. A better option is to return the status that checks if the target file exists locally (succeeded to download eventually).

The reason we download ClamAV is because OneBranch fails to do so. When OneBranch fails to install ClamAV, it sends a warning, not an error. Using `continueOnError: true` is a better option for this case, as we can notice the step failed, but doesn't stop the rest of the pipeline.

Related work items: #7168