Merged PR 18390: docs: Adding documentation for running tests on Baremetal - Part 1
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->

This PR is the initial documentation for the running of tests on Baremetal Hosts. This PR provides a flowchart to give an understanding of logical flow and what code resides where.

There will be a following PR giving details of the BM Server like IPs, Proxy used, HTTPD Server being used and so on, when the PR for running tests on Supermicros are ready to merge. Since at this point of time, there is only a single flavor of BM, I have limited information to code flow.

Seeing the flowcharts directly on Azure Devops is blocked by the following - https://developercommunity.visualstudio.com/t/Support-mermaid-diagrams-in-markdown-fil/817244

The mermaid flowcharts can be seen locally on VSCode by adding the mermaid extension to VSCode.- https://marketplace.visualstudio.com/items?itemName=vstirbu.vscode-mermaid-preview

The reason to use mermaid is that it gives clean flowcharts and is easy to use to create complicated diagrams.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

Documentation to show how tests run on BM Server.

# 📝 Checks
<!-- Note any manual checks you've run, base list of tests: dev-docs/manual-validation.md -->

N/A

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

I'll add more documentation when the Supermicros PR is ready - https://dev.azure.com/mariner-org/ECF/_git/trident/pullrequest/18139

# 🗒️ Notes
<!-- Add any other relevant info -->

Related work items: #7153