Merged PR 18434: feature: Basic Netplan Validation
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->

Check for basic errors in netplan.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

Check basic errors on validation to reduce the likelihood of booting to a netplan error.

Related work items: #6615