Merged PR 18474: feature: Simplify runtime container
# 🔍 Description
Runtime container should use RPMs produced as part build the build, instead of rebuilding. As an added benefit, rebuilding the runtime container is faster now, as it caches the dependencies layer.

# 🤔 Rationale
Prepare container definition for possible release and for e2e testing. Ensure it is using the same RPM bits, that would be subsequently released.

Related work items: #6650