Merged PR 18497: engineering: Don't build trident within container when producing RPM locally
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description

This changes the Makefile to build trident outside of docker, and then copy the fully built binary in. This is useful because it means that we can rely on Cargo's change detection to decide whether the trident binary needs to be rebuild, and only rebuild the RPM if so.

# 🤔 Rationale

Always rebuilding the RPM is unfortunate because it means that we'll constantly have to rebuild the ISO every time we `make run-netlaunch`. This PR drastically speeds up cases of re-running netlaunch with only the HC changed.

Related work items: #7277