Merged PR 18499: engineering: Add support for containerhost test image
# 🔍 Description
Adds logic to generate containerhost MOS and runtime image so that we can test trident e2e from container. This is only a starting point, E2E tests need to be switched to using MIC ISO before this can be integrated.

# 🤔 Rationale
As customers are using Trident from docker container, we need to have e2e tests for that scenario.

Related work items: #6650