Merged PR 18566: engineering: Staging pipeline refactor
# 🔍 Description

Build staging:

| Pipeline | Source | Publishes to | Testing |
| ---------| -------- | ------------- | -------- |
| PR | Build | Nowhere | Check, UTs, Build |
| ci | Build | `rpms-ci` | Check, UTs, Build  & E2E |
| prerelease | `rpms-ci` | `rpms-prerelease` |  E2E & Baremetal |
| release (new*) | `rpms-prerelease` | `rpms` | TBD |

\* To be added in follow-up PR.

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->

Provide clear and reliable staging with explicit expectations for each stage.

# Sample

Ran a tweaked version of this to test BM test yamls:
https://dev.azure.com/mariner-org/ECF/_build/results?buildId=542234&view=results

![image.png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/18566/attachments/image.png)
(Bc this is just a pipeline change, getting one to run seems validation enough)

Related work items: #7182, #7184