Merged PR 18603: engineering: update instructions on running from a container
# 🔍 Description
Update the docs to bind entire /run for trident running from container

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->
We have been having issues with RAID running from the container. mdadm requires `/run/mdadm`.
We have decided to bind entire `/run`

# 📝 Checks
<!-- Note any manual checks you've run, base list of tests: dev-docs/manual-validation.md -->

# 📌 Follow-ups
<!-- When applicable, add links to tasks or PR for work that will be addressed
later as a follow up of this PR. -->

- #0000

# 🗒️ Notes
<!-- Add any other relevant info -->

engineering: update instructions on running from a container

Related work items: #6850