Merged PR 18563: engineering: Implement A/B update in E2E testing
<!-- Please, remember to prefix the title with one of:
 - `bug:`
 - `feature:`
 - `engineering:`
 - `docs:`.
For details, check [Pull Requests](/dev-docs/pull-requests.md) and
[this page](https://dev.azure.com/mariner-org/ECF/_git/ecf-docs?version=GBmain&path=/TeamDocs/dev-guide.md&_a=preview). -->

# 🔍 Description
<!--
What is this PR about?
Reviewers should get a clear understanding of what is being done/changed.
-->
This PR implements the basic logic for integrating A/B updates into E2E testing of trident. Specifically, e2e-test-run.yml now contains the logic to parse the Trident config, determine if A/B update testing is required, and invoke ab_update.py, a script that triggers A/B update on the host, by updating the local Trident config and re-running Trident.

Currently, the tasks are not executed for the verity images: this is because there seems to be a bug where post A/B update, we cannot always re-connect to the host via ssh. In the next PR, this will be fixed to also conduct A/B update testing on Trident configs that require verity images. Similarly, A/B testing will be extended to the container images in a future PR.

In this first version, two A/B updates are performed per Trident config that contains abUpdate: prov OS -> runtime OS A -> runtime OS B -> runtime OS A. The update images are pulled from the published artifacts and then renamed/version to "simulate" updated images. In the future, this logic will be expanded to perform 5 A/B updates per Trident config, with 5 different images pulled from http or Azure blob storage and/or artifacts (TBD).

# 🤔 Rationale
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->
This change is necessary so that every change to the Trident code base can be tested in the A/B update flow, in an automated fashion.

#Owners
<!-- (can be skipped for bug fixes)
Why is this PR needed?
Reviewers should get a clear understanding of the goals of this PR.
-->
Himaja Kesari himajakesari@microsoft.com
Nadiia Dubchak ndubchak@microsoft.com

Related work items: #6559