Merged PR 18714: bug: Set job timeout for Testing back to 25 min and add individual timeouts for A/B testing
# 🔍 Description

This PR sets the job-wide timeout for Testing back to 25 min and adds individual timeouts for A/B testing steps.

# 🤔 Rationale

Currently, there is a single timeoutInMinutes for the entire Testing job in netlaunch-testing.yml. As a result, if a task times out, e.g. because of the ssh issue that is now being investigated, the remaining steps, mostly Publish OS disk, are not executed. As a result, we do not publish the qcow2 image on failures that are caused by timeouts. This PR fixes that by setting the job-wide timeout back to 25 min and adding timeouts for individual tasks related to A/B updates.

# 📝 Checks

- [X] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

Related work items: #7508