Merged PR 18643: engineering: Implement tracestream for Trident to send telemetry events to netlaunch
# 🔍 Description

What is this PR about?

We need Trident to be able to stream telemetry events to netlaunch, so implementing tracestream (inspired by logstream) is a way to accomplish this. This PR is attempting to introduce tracestream and have it work as a custom layer to the tracing crate.

# 🤔 Rationale

Why is this PR needed?

We want telemetry events streamed to netlaunch so we can upload to Kusto DB. This PR introduces the framework to allow for performance events to be created within Trident.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:
- #6847
- #7480
- #7481
- #7482
- #7483

# 🗒️ Notes

Initial implementation of tracestream

Related work items: #7381