Merged PR 18724: bug: Fix A/B update logic in BM E2E testing
# 🔍 Description

This PR fixes a bug where the Trident images were not found at the expected paths in BM testing. This PR adds a step to rename images to fit the logic in e2e-test-run.yml. Moreover, this PR changes the step to check SSH connection to use exponential backoff and have a longer total wait time.

# 🤔 Rationale

This PR is needed to enable A/B update testing for BM.

----
#### AI description (iteration 1)
#### PR Classification
This pull request is a bug fix, addressing the issue of incorrect naming of Trident images used for A/B update end-to-end (E2E) testing.

#### PR Summary
The pull request modifies the deployment testing pipeline to correctly rename the Trident and Verity images used for functional and E2E testing.
- In the file `/.pipelines/templates/stages/deployment_testing/baremetal/baremetal-prep.yml`, code has been added to rename the Trident and Verity images, ensuring they are correctly identified during testing.

Related work items: #7520, #7539