Merged PR 18680: engineering: Add execroot bind bound and use it for datastore
# 🔍 Description

This creates /tmp/execroot that points to the root of the system Trident was run from, and then switches the datastore to use this bind mount for accessing the datastore within the chroot.

# 🤔 Rationale

This enables access to files on the system trident was run from even while within the chroot.

Related work items: #6623, #6624, #7339