Merged PR 18808: engineering: Minor fixes to functional tests
# 🔍 Description

Functional tests were not copying over the trident and osmodifier binaries to the netlaunch serve path to run correctly.

Also, because we no longer use isogen and rebuild the ISO with trident inside we can get rid of the logic to regenerate it inside the functional tests, we can instead rely on make to do this.

----
#### AI description (iteration 1)
#### PR Classification
This pull request primarily focuses on minor fixes and improvements to the functional tests in the software project.

#### PR Summary
The changes in this pull request streamline the functional testing process by modifying the Makefile and several Python scripts. The modifications also include the removal of some unnecessary code.
- In the `/Makefile`, the `functional-test` target has been updated to include `bin/netlaunch` and `bin/trident-mos.iso` in the prerequisites, and the `ARGUS_TOOLKIT_PATH` variable has been removed.
- In `/functional_tests/test_setup.py`, the `deploy_vm` function no longer requires the `installer_iso_path` parameter, and the `INSTALLER_ISO_PATH` and `NETLAUNCH_SERVE_DIRECTORY` constants have been introduced.
- In `/functional_tests/conftest.py`, the `INSTALLER_ISO_PATH` and `NETLAUNCH_SERVE_DIRECTORY` constants have been defined.
- In `/trident-mos/files/root.profile`, the alias for viewing trident's logs has been removed and the `journalctl` command has been updated.

Related work items: #7584