Merged PR 18818: engineering: BuildNumber unchanged by rerun to always match trident version
# 🔍 Description
 Now get-version.py returns the BuildNumber if the BuildNumber already is the trident version.

# 🤔 Rationale
get-version.py produced bad output on re-run. Jobs can fail by multiple reasons and rerunning the failed jobs is a common practice. For these cases BuildNumber should keep being the correct one.

Why is this PR needed?
When rerunning the pipeline (build trident rpms stage) the BuildNumber has already been change to the trident version, so the wrong input was given to get-version.py file. This produces the wrong output, which isn't compatible as an id to publish in the universal feed. Now, if the BuildNumber matches with the trident version format, it will return the BuildNumber as it already has been changed to be the trident version.

# 📝 Checks

In this Pipeline run, it can be seen that in the rerun the `get-version.py` script returns the BuildNumber as it already matches to the trident version: https://dev.azure.com/mariner-org/ECF/_build/results?buildId=557115&view=logs&j=827c145e-7255-57e9-fe55-91f584068c21&t=1f5fdca5-4059-51b0-056b-2ac5af789efb&l=12

----
#### AI description (iteration 1)
#### PR Classification
This pull request is a bug fix for the `get-version.py` script to ensure the BuildNumber remains consistent with the trident version even on reruns.

#### PR Summary
The changes in this pull request address a bug in the `get-version.py` script that was causing the trident-ci to fail due to incorrect output on re-run. The fix ensures that the BuildNumber matches the trident version consistently.
- In `/scripts/get-version.py`, the logic for generating the version number has been updated. Now, if the BuildNumber already matches the trident version, it is directly printed. Otherwise, the previous logic of generating the version number from the date and id is followed. This change ensures that the BuildNumber remains consistent across reruns.

Related work items: #7587