Merged PR 18820: engineering: Use a pinned version of baremetal.vhdx in Makefile
# 🔍 Description

This way we can make sure we're using a consistent version of the base image

----
#### AI description (iteration 1)
#### PR Classification
This pull request is a code cleanup that pins the version of `baremetal.vhdx` used in the Makefile.

#### PR Summary
The pull request modifies the Makefile to use a specific version of `baremetal.vhdx` instead of any available version. This change ensures consistency and predictability in the build process.
- In the Makefile, the `BASE_IMAGE_VERSION` variable has been updated from a wildcard (`*`) to a specific version (`2.0.20240403`).

Related work items: #7605