Merged PR 18819: feature: Use OS Modifier from exec root
# 🔍 Description

Use OS modifier directly from the filesystem Trident was executed from

# 🤔 Rationale

Now that we have the exec root bind mounted during the configure step, we don't have to worry about copying OS modifier onto the runtime OS.

----
#### AI description (iteration 1)
#### PR Classification
This pull request introduces a new feature that modifies the OS configuration module to use the OS Modifier from the execution root.

#### PR Summary
The changes in this pull request primarily revolve around the modification of the OS configuration module to use the OS Modifier from the execution root, instead of the previously hardcoded path. This change enhances the flexibility and adaptability of the system.
- In `src/modules/osconfig/mod.rs`, the `provision` function was removed and the `configure` function was updated to use the OS Modifier from the execution root.
- Changes in `src/modules/hooks.rs` and `src/modules/mod.rs` reflect the updated function signatures and the new approach to using the OS Modifier.
- The `set_up_users` and `set_up_hostname` functions in `src/modules/osconfig/users.rs` and `src/modules/osconfig/hostname.rs` respectively, were updated to use the OS Modifier from the provided path.
- Other modules such as `src/modules/boot/mod.rs`, `src/modules/management.rs`, `src/modules/network/mod.rs`, and `src/modules/storage/mod.rs` were updated to accommodate the new function signature.

Related work items: #6751