Merged PR 18849: engineering: Compress VM image before publishing
# 🔍 Description

When a E2E test fails we capture a VM image for debugging. This changes the task to zstd compress the image prior to publishing as a pipeline artifact.

# 🤔 Rationale

The VM image is ~11 GB but after compression should be closer to 500 MB because the disk is mostly empty

----
#### AI description (iteration 1)
#### PR Classification
This pull request is focused on optimizing the VM image publishing process by introducing compression.

#### PR Summary
The changes in this pull request aim to compress the VM image before publishing, reducing the size and potentially speeding up the transfer process.
- In `/.pipelines/templates/stages/deployment_testing/vm/netlaunch-prep.yml`, the `zstd` compression library was added to the list of dependencies.
- In `/.pipelines/templates/stages/deployment_testing/vm/netlaunch-testing.yml`, a new command was introduced to compress the VM image using `zstd` before it is published.

Related work items: #7618