Merged PR 18868: engineering: Retry Universal Package Downloads
# 🔍 Description

Extension of !18847 for universal package downloads, which sometimes have network errors.

----
#### AI description (iteration 1)
#### PR Classification
This pull request introduces a feature to retry Universal Package downloads in the pipeline tasks.

#### PR Summary
The changes in this pull request aim to improve the reliability of package downloads during pipeline execution by adding a retry mechanism. This is achieved by modifying the pipeline templates to include a retry count on task failure.
- In `/.pipelines/templates/stages/common_tasks/download-emu.yml`, a `retryCountOnTaskFailure` parameter was added to the UniversalPackages task with a value of 3, allowing the task to retry thrice upon failure.
- Similar changes were made in `/.pipelines/templates/stages/common_tasks/download-universal-artifact.yml`, adding a `retryCountOnTaskFailure` parameter to the UniversalPackages task.
- Minor formatting changes were also made in both files, replacing single quotes with double quotes for consistency.

Related work items: #7616