Merged PR 18836: engineering: Make /run a tmpfs in chroot and remove /run partitions from E2E tests
# 🔍 Description

The /run directory is supposed to be a temporary filesystem, but some of our E2E tests were creating an actual partition for it.

----
#### AI description (iteration 1)
#### PR Classification
This pull request is a code modification aimed at improving the system's file management by making `/run` a tmpfs in chroot and removing `/run` partitions from E2E tests.

#### PR Summary
The changes in this pull request primarily focus on modifying the handling of the `/run` directory in the system and updating the configurations in the E2E tests.
- In `src/modules/mod.rs`, a new tmpfs mount for the `/run` directory is created during the initialization of a new root.
- The configurations for the E2E tests in `trident_configurations` (abupdate-with-verity, combined, rerun, verity) have been updated to remove the `/run` partitions, reflecting the changes in the system's file management.

Related work items: #7609