Merged PR 18891: engineering: Changing openssl for ssh-keygen
# 🔍 Description
Changing the way we adjust the format of the generated ssh-keys. Using ssh-keygen instead of openssl.

# 🤔 Rationale

ssh-keygen is more consistent throw versions than openssl. This was currently and issue as the agents that run on BM were using ubuntu 20 and for the VM testing the pipeline was using ubuntu 22, both with a different version of openssl with different behavior.

# 📝 Checks
Pipeline runs:
- https://dev.azure.com/mariner-org/ECF/_build/results?buildId=560694&view=results
- https://dev.azure.com/mariner-org/ECF/_build/results?buildId=560709&view=results

----
#### AI description (iteration 1)
#### PR Classification
This pull request is a code modification aimed at resolving a failing test in the Baremetal Deployment Testing stage.

#### PR Summary
The pull request modifies the `trident-prep.yml` file to replace the use of `openssl` with `ssh-keygen` for key generation in the SSH setup for testing. This change is expected to address the linked work item, which reported failures in the Baremetal Deployment Testing stage.
- Key change in `/.pipelines/templates/stages/deployment_testing/common_tasks/trident-prep.yml`: Replaced `openssl` command with `ssh-keygen` for SSH key generation.

Related work items: #7636