Merged PR 18896: engineering: Update baremetal base image
# 🔍 Description

Use the newer baremetal base image. Run `make -B artifacts/baremetal.vhdx` to update

----
#### AI description (iteration 1)
#### PR Classification
This pull request updates the version of the base image used in the baremetal build process.

#### PR Summary
The pull request modifies the Makefile to use a newer version of the base image for the baremetal build. This change could potentially improve the build process and the resulting product.
- Updated the `BASE_IMAGE_VERSION` in the Makefile from `2.0.20240403` to `2.0.20240425`.

Related work items: #7674